use more complex workflow
This workflow has two key differences to the previous, simple one:

- it tests lots of different Perl versions on each platform, not just
  5.32
- it splits building and testing into different steps that can fail
  independently. The build process does not need to run on all
platforms. It's enough to run on the one that you would usually use (or
typically, Linux). Author tests don't need to run on the other
platforms.

It also adds support for measuring code coverage, which will show up in
PRs as a metric with changes highlighted. To make code coverage work you
will have to create an account at codecov.io and sync the repository. This'll
allow you to see changes to code coverage on PRs.